adds separate place search routes extension, replaces load place with reviews
