+++ Daydream/Networking/API+PlaceSearch.swift
@@ -13,58 +13,10 @@ import GooglePlaces
 
 extension API {
     enum PlaceSearch {
-        struct TextSearch {
-            enum QueryType {
-                case touristSpots, restaurants, placeByName
-            }
-
-            private let baseUrlString = "https://maps.googleapis.com/maps/api/place/textsearch/json?query="
-
-            let url: URL
-
-            init?(name: String, location: CLLocationCoordinate2D? = nil, queryType: QueryType) {
-                guard let keyParam = AppDelegate.getAPIKeys()?.googleAPI else {
-                    return nil
-                }
-                // e.g., İstanbul -> Istanbul
-                let placeNameStripped = name.folding(options: .diacriticInsensitive, locale: .current)
-                let placeWords = placeNameStripped.split(separator: " ")
-                var queryParam = ""
-
-                switch queryType {
-                case .placeByName:
-                    break
-                case .touristSpots:
-                    queryParam = "tourist+spots+in"
-                case .restaurants:
-                    queryParam = "restaurants+in"
-                }
-
-                placeWords.forEach { word in
-                    queryParam += "+" + word
-                }
-
-                let urlString: String
-
-                if let location = location {
-                    let locationParam = "\(location.latitude),\(location.longitude)"
-                    let radiusParam = "10000"
-                    urlString = "\(baseUrlString)\(queryParam)&location=\(locationParam)&radius=\(radiusParam)&key=\(keyParam)"
-                } else {
-                    urlString = "\(baseUrlString)\(queryParam)&key=\(keyParam)"
-                }
-
-                guard let url = URL(string: urlString) else {
-                    return nil
-                }
-                self.url = url
-            }
-        }
-
         /// Can be used to return multiple Google Place objects (e.g., sights, fallback restaurants) based on the city name, optional location,
         /// and query type.
-        static func loadPlaces(name: String, location: CLLocationCoordinate2D? = nil, queryType: TextSearch.QueryType) -> AnyPublisher<[Place], Error>? {
-            guard let url = TextSearch(name: name, location: location, queryType: queryType)?.url else {
+        static func loadPlaces(name: String, location: CLLocationCoordinate2D? = nil, queryType: TextSearchRoute.QueryType) -> AnyPublisher<[Place], Error>? {
+            guard let url = TextSearchRoute(name: name, location: location, queryType: queryType)?.url else {
                 return nil
             }
             return URLSession.shared.dataTaskPublisher(for: url)
@@ -77,7 +29,7 @@ extension API {
         /// Convenience method to return one Google place. Uses loadPlaces(name:location:queryType:) and ensures the first element is returned.
         static func loadPlace(name: String,
                               location: CLLocationCoordinate2D? = nil,
-                              queryType: TextSearch.QueryType,
+                              queryType: TextSearchRoute.QueryType,
                               receiveOnMainQueue: Bool = true) -> AnyPublisher<Place, Error>? {
             let publisher = PlaceSearch.loadPlaces(name: name, location: location, queryType: queryType)?
                 .tryMap { places -> Place in
@@ -92,6 +44,24 @@ extension API {
             } else {
                 return publisher?.eraseToAnyPublisher()
             }
+        }
+
+        /// Returns a Place object with at least one review.
+        static func loadPlaceWithReviews(placeId: String) -> AnyPublisher<Place, Error>? {
+            guard let url = PlaceDetailsRoute(placeId: placeId)?.url else {
+                return nil
+            }
+            return URLSession.shared.dataTaskPublisher(for: url)
+                .map { $0.data }
+                .decode(type: PlaceCollection.self, decoder: JSONCustomDecoder())
+                .tryMap { collection -> Place in
+                    guard !collection.result.reviews.isEmpty else {
+                        throw NetworkError.insufficientResults
+                    }
+                    return collection.result
+                }
+                .receive(on: DispatchQueue.main)
+                .eraseToAnyPublisher()
         }
 
         /// Load photo as UIImage using Google Places SDK
