+++ Daydream/Networking/API.swift
@@ -10,5 +10,5 @@ import Foundation
 import Combine
 import CoreLocation
 
-// Creates namespace for API calls
+// Creates namespace for API calls. Routes for each namespaced API are in extensions.
 enum API {}
